WIP: session: Ignore DetachFromTarget error if target is gone (#110)
rpcc.Conn.Close calls Target.DetachFromTarget before closing the
websocket connection, which fails if the target is already closed.
This is a bit inconvenient because we have to ignore errors from
rpcc.Conn.Close when we close the target in advance, which means that
we might miss other interesting errors.

This change updates the detach function to ignore errors on
Target.DetachFromTarget calls if it is because the target is already
gone.

Also adds a unit test to test this behavior.

FIXME: The unit test still does not pass.